requirements.txt
pandas>=2.0
numpy>=1.24
python-dotenv>=1.0
pydantic>=2.0
scikit-learn>=1.3
google-genai>=0.3.0
tqdm>=4.66
streamlit>=1.31
